+++ packages/i18n/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.66](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.65...@stream-io/i18n-0.0.1-alpha.66) (2023-05-25)
+
+
+
+## [0.0.1-alpha.65](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.64...@stream-io/i18n-0.0.1-alpha.65) (2023-05-25)
+
+
+
+## [0.0.1-alpha.64](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.63...@stream-io/i18n-0.0.1-alpha.64) (2023-05-25)
+
+
+
+## [0.0.1-alpha.63](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.62...@stream-io/i18n-0.0.1-alpha.63) (2023-05-25)
+
+
+
+## [0.0.1-alpha.62](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.61...@stream-io/i18n-0.0.1-alpha.62) (2023-05-25)
+
+
+
+## [0.0.1-alpha.61](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.60...@stream-io/i18n-0.0.1-alpha.61) (2023-05-25)
+
+
+
+## [0.0.1-alpha.60](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.59...@stream-io/i18n-0.0.1-alpha.60) (2023-05-25)
+
+
+
 ## [0.0.1-alpha.59](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.58...@stream-io/i18n-0.0.1-alpha.59) (2023-05-25)